feat: 전체 인증 삭제 가능
