+++ backend/app/src/main/java/com/yagubogu/checkin/controller/v1/CheckInController.java
@@ -15,6 +15,7 @@ import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -33,6 +34,16 @@ public class CheckInController implements CheckInControllerInterface {
         checkInService.createCheckIn(memberClaims.id(), request);
 
         return ResponseEntity.status(HttpStatus.CREATED).build();
+    }
+
+    @Override
+    public ResponseEntity<Void> deleteCheckIn(
+            final MemberClaims memberClaims,
+            @PathVariable final Long checkInId
+    ) {
+        checkInService.deleteCheckIn(memberClaims.id(), checkInId);
+
+        return ResponseEntity.noContent().build();
     }
 
     public ResponseEntity<CheckInCountsResponse> findCheckInCounts(
