+++ backend/app/src/main/java/com/yagubogu/checkin/controller/v1/CheckInControllerInterface.java
@@ -16,7 +16,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.time.LocalDate;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -35,6 +37,18 @@ public interface CheckInControllerInterface {
     ResponseEntity<Void> createCheckIn(
             @Parameter(hidden = true) MemberClaims memberClaims,
             @RequestBody CreateCheckInRequest request
+    );
+
+    @Operation(summary = "경기 인증 삭제", description = "지정한 경기 인증을 삭제합니다. 본인의 인증만 삭제할 수 있습니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "204", description = "인증 삭제 성공"),
+            @ApiResponse(responseCode = "403", description = "본인의 체크인만 삭제할 수 있음"),
+            @ApiResponse(responseCode = "404", description = "체크인을 찾을 수 없음")
+    })
+    @DeleteMapping("/{checkInId}")
+    ResponseEntity<Void> deleteCheckIn(
+            @Parameter(hidden = true) MemberClaims memberClaims,
+            @PathVariable Long checkInId
     );
 
     @Operation(summary = "인증 수 조회", description = "연도별로 회원의 총 인증 수를 조회합니다.")
